+++ src/map_utils.cpp
@@ -16,11 +16,16 @@
  * You should have received a copy of the GNU General Public License
  * along with this program; if not, write to the Free Software
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
- *
  */
+
+
+
 
 #include <cmath>
 #include <cstdlib>
+
+
+
 
 #include "map_utils.h"
 #include "globals.h"
@@ -86,7 +91,7 @@ int SlavGPS::map_utils_mpp_to_scale(double mpp)
  * Returns: a Zoom Level.
  * See: http://wiki.openstreetmap.org/wiki/Zoom_levels
  */
-uint8_t SlavGPS::map_utils_mpp_to_zoom_level(double mpp)
+int SlavGPS::map_utils_mpp_to_zoom_level(double mpp)
 {
 	int answer = 17 - map_utils_mpp_to_scale(mpp);
 	if (answer < 0) {
